+++ libs/data-mapper-v2/src/components/schema/tree/styles.ts
@@ -97,25 +97,23 @@ export const useHandleStyles = makeStyles({
   connectedAndSelected: {
     backgroundColor: colors.handleActive,
   },
-  repeatingNode: {
-    height: '16px',
-    width: '16px',
+  repeating: {
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
+    height: '16px',
+    width: '16px',
   },
-  repeatingConnection: {
-    height: '16px',
-    width: '16px',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
+  repeatingIcon: {
+    fontSize: '12px',
+    pointerEvents: 'none',
+    color: '#ddd',
   },
   repeatingConnectionIcon: {
     color: 'white',
     fontSize: '12px',
   },
-  repeatingNodeIcon: {
+  repeatingAndActiveNodeIcon: {
     color: colors.handleActive,
     fontSize: '12px',
   },