+++ libs/vscode-extension/vitest.config.ts
@@ -6,8 +6,6 @@ export default defineProject({
   plugins: [react()],
   test: {
     name: packageJson.name,
-    dir: './src',
-    watch: false,
     environment: 'jsdom',
     setupFiles: ['test-setup.ts'],
     coverage: { enabled: true, provider: 'istanbul', include: ['src/**/*'], reporter: ['html', 'cobertura'] },